Refactor project structure into two separate projects
The front-end and back-end were previously stored in one repository for simple storage on Github, so that all the code could be seen easily. However, this makes the project harder to run locally since the front-end and back-end run on different ports. The projects are therefore being separated into two different repositories on Github.